% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcentral-package.R
\docType{package}
\name{sfcentral-package}
\alias{sfcentral}
\alias{sfcentral-package}
\title{sfcentral: Spatial Centrality and Dispersion Statistics}
\description{
Computing centrographic statistics (central points, standard distance, standard deviation ellipse, standard deviation box) for observations taken at point locations in 2D or 3D. The 'sfcentral' library was inspired in 'aspace' package but conceived to be used in a spatial 'tidyverse' context.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gavg712.gitlab.io/sfcentral/}
  \item \url{https://gitlab.com/gavg712/sfcentral}
  \item Report bugs at \url{https://gitlab.com/gavg712/sfcentral/-/issues}
}

}
\author{
\strong{Maintainer}: Gabriel V. Gaona \email{gabo@gavg712.com} (\href{https://orcid.org/0000-0003-4666-2248}{ORCID})

}
\keyword{internal}
